requirements.txt
blinker==1.8.2
camelot-py==0.11.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==3.0.3
GeoAlchemy2==0.15.2
geopandas==1.0.1
ghostscript==0.7
greenlet==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20240706
plotly==5.23.0
psycopg2-binary==2.9.9
pycparser==2.22
pyogrio==0.9.0
pypdf==4.3.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
scipy==1.14.1
setuptools==72.1.0
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.31
tabulate==0.9.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
xlrd==2.0.1
zipp==3.20.2
